Prevent Last API Key Deletion

+++ Modules/User/Http/Controllers/Admin/Account/ApiKeysController.php
@@ -41,6 +41,11 @@ class ApiKeysController extends AdminBaseController
 
     public function destroy(UserToken $userToken)
     {
+        if ($this->userToken->allForUser($this->auth->id())->count() === 1) {
+            return redirect()->route('admin.account.api.index')
+                ->withFail(trans('user::users.last token can not be deleted'));
+        }
+
         $this->userToken->destroy($userToken);
 
         return redirect()->route('admin.account.api.index')
